Move grid plotting utilities to navigation.plotting and fix ActiveSkill import
make_plotting_grid and plot_grid are general grid utilities, not
ProcTHOR-specific. Move them to railroad.navigation.plotting and
delete the now-empty procthor/plotting.py.

Also fix skill/__init__.py to import ActiveSkill from environment.py
where it now lives.
