+++ packages/railroad/src/railroad/navigation/__init__.py
@@ -1,6 +1,6 @@
 """Reusable grid navigation primitives (pathing, occupancy grid mixin)."""
 
-from . import pathing
+from . import pathing, plotting
 from .constants import COLLISION_VAL, FREE_VAL, OBSTACLE_THRESHOLD, UNOBSERVED_VAL
 from .occupancy_grid_mixin import OccupancyGridPathingMixin
 
@@ -11,4 +11,5 @@ __all__ = [
     "OccupancyGridPathingMixin",
     "UNOBSERVED_VAL",
     "pathing",
+    "plotting",
 ]
